+++ .env
@@ -1,5 +1,5 @@
 # version
-GOLANG_VERSION=1.19.3
+GOLANG_VERSION=1.21
 K6_VERSION=0.42.0
 
 # service
